requirements.txt
Django==4.2.7
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
